Add pages/_.vue for dynamically routing to CMS pages
When a user navigates to a page not otherwise handled by static
routes, this page will attempt to find a document in the nuxt
content directory at the same relative path. If found, it renders
the document; if not, it returns a 404 status.

Later commits will style and improve the visuals of this page.